se insertan en la bdd todo menos los sponsors

+++ ada_gonzalez/LNHH/src/main/java/dao/DAO.java
@@ -6,7 +6,7 @@ public interface DAO <T>{
 	public void insert(T entity);
 	public void delete(T entity);
 	public List<T> selectAll();
-	public T select(int id);
+	public T select(long id);
 	public void update(T entity);
 	
 }
